+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -1808,7 +1808,7 @@ class RunAgentAgentToolInputRunTools(BaseModel):
 
     schema_: Annotated[AgentToolInputRunSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KPNT9J7SB134GT6H353VVAH6"
+    id: Optional[str] = "01KPPVWA00TWYC38JRNXZZHEY4"
 
     description: Optional[str] = None