+++ packages/orq-rc/RELEASES.md
@@ -7109,3 +7109,13 @@ Based on:
 - [python v4.8.0-rc.33] packages/orq-rc
 ### Releases
 - [PyPI v4.8.0-rc.33] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.33 - packages/orq-rc
+
+## 2026-04-21 01:56:09
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.8.0-rc.34] packages/orq-rc
+### Releases
+- [PyPI v4.8.0-rc.34] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.34 - packages/orq-rc